% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Chen_thermal_traits_meta}
\alias{Chen_thermal_traits_meta}
\title{Chen (2015) Thermal Traits metadata}
\format{\preformatted{Classes 'tbl_df', 'tbl' and 'data.frame':	21 obs. of  3 variables:
 $ Variable   : chr  "References" "ID" "Phylum" "Class" ...
 $ Description: chr  "References" "ID" "Phylum" "Class" ...
 $ Units      : chr  NA NA NA NA ...
}}
\source{
\url{http://plankt.oxfordjournals.org/content/37/2/285/suppl/DC1}
}
\usage{
data(Chen_thermal_traits_meta)
}
\description{
Metadata for data from Chen B. (2015) Patterns of thermal limits of phytoplankton.
Journal of Plankton Research 37, 285–292.
}
\keyword{datasets}

