% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Chen_thermal_traits_references}
\alias{Chen_thermal_traits_references}
\title{Chen (2015) Thermal Traits references.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 1 columns.}
\source{
\url{http://plankt.oxfordjournals.org/content/37/2/285/suppl/DC1}
}
\usage{
data(Chen_thermal_traits_references)
}
\description{
References for data from Chen B. (2015) Patterns of thermal limits of phytoplankton. 
Journal of Plankton Research 37, 285–292.
}
\keyword{datasets}

