% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rd_table.R
\name{create_rd_table}
\alias{create_rd_table}
\title{create_rd_table}
\usage{
cat(tabular(df))
}
\arguments{
\item{df}{a dataframe}
}
\value{
an rd compatible dataframe
}
\description{
Creates an .Rd compatible dataframe. Adapted from the suggested
 function in the roxygen2 vignette 
 https://cran.r-project.org/web/packages/roxygen2/vignettes/formatting.html
}

