% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Edwards_cell_volumes}
\alias{Edwards_cell_volumes}
\title{Cell volumes of phytoplankton.}
\format{A data frame with 134 rows and 4 variables:

\bold{Encoding }
\encoding{UTF-8}
\tabular{llll}{
\bold{Column name} \tab \bold{Variable definition} \tab \bold{Units} \tab \bold{Storage type} \cr
species         \tab Species name                    \tab text  \tab string        \cr
isolate         \tab Isolate ID                      \tab text  \tab string        \cr
volume          \tab Cell volume                     \tab µm^3^ \tab floating point\cr
volume_citation \tab Citation for volume measurement \tab text  \tab string        
}}
\source{
\url{http://www.esapubs.org/archive/ecol/E096/202/}
}
\usage{
data(Edwards_cell_volumes)
}
\description{
Data from Kyle F. Edwards, Christopher A. Klausmeier, and Elena Litchman (2015). 
Nutrient utilization traits of phytoplankton. Ecology 96:2311. 
http://dx.doi.org/10.1890/14-2252.1.
}
\details{
Table 3 of Ecological Archives E096-202-D1.
}
\keyword{datasets}

