% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Edwards_citations}
\alias{Edwards_citations}
\title{Citation information for nutrient utilization traits.}
\format{\preformatted{Classes 'tbl_df', 'tbl' and 'data.frame':	137 obs. of  2 variables:
 $ citation_number: int  1 2 3 4 5 6 7 8 9 10 ...
 $ full_citation  : chr  "Admiraal, W. and D. Werner. 1983. Utilization of limiting concentrations of ortho-phosphate and production of extracellular org"| __truncated__ "Ahlgren, G. 1977. Growth of Oscillatoria agardhii in chemostat culture. 1. Nitrogen and phosphorus requirements. Oikos 29:209-2"| __truncated__ "Ahlgren, G. 1978. Growth of Oscillatoria agardhii in chemostat culture. 2. Dependence of growth constants on temperature. Mitt."| __truncated__ "Ahlgren, G. 1985. Growth of Oscillatoria agardhii in chemostat culture. 3. Simultaneous limitation of nitrogen and phosphorus. "| __truncated__ ...
}}
\source{
\url{http://www.esapubs.org/archive/ecol/E096/202/}
}
\usage{
data(Edwards_citations)
}
\description{
Data from Kyle F. Edwards, Christopher A. Klausmeier, and Elena Litchman (2015).
Nutrient utilization traits of phytoplankton. Ecology 96:2311.
http://dx.doi.org/10.1890/14-2252.1.
}
\details{
Table 2 of Ecological Archives E096-202-D1.
}
\keyword{datasets}

