% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Kremer_raw}
\alias{Kremer_raw}
\title{Phytoplankton cell and natural unit biovolumes. 
Raw biovolume data - pre quality control.}
\format{A data frame with 312238 rows and 18 variables:

\bold{Encoding }
\encoding{UTF-8}
\tabular{lll}{
\bold{Variable} \tab \bold{Definition} \tab \bold{Units} \cr
nu.measurement.ID  \tab Index for individual estimate of a natural unit's biovolume                                     \tab -                     \cr
data.source        \tab Source of biovolume data                                                                        \tab -                     \cr
location           \tab Sampling location as reported in original data source                                           \tab -                     \cr
sample.date        \tab Date of biovolume sample, from original data source (when reported)                             \tab Days                  \cr
source.ID          \tab Sample ID from original data source                                                             \tab -                     \cr
source.taxa.name   \tab Taxa name reported in source data.                                                              \tab -                     \cr
source.taxa.ID     \tab Taxonomic ID code from original data source                                                     \tab -                     \cr
original.taxa.name \tab Taxa name, combining genus or genus + species names from original data source.                  \tab -                     \cr
genus              \tab Genus name                                                                                      \tab -                     \cr
species            \tab Species name                                                                                    \tab -                     \cr
subspecies         \tab Subspecies name                                                                                 \tab -                     \cr
variety            \tab Variety name                                                                                    \tab -                     \cr
form               \tab Form name                                                                                       \tab -                     \cr
nu                 \tab Descriptor of natural unit type                                                                 \tab -                     \cr
cells.per.nu       \tab Cells per natural unit                                                                          \tab Cells per natural unit\cr
nu.biovol          \tab Biovolume corresponding to a natural unit of the specified taxa                                 \tab log10(μm^3^)         \cr
cell.biovol        \tab Biovolume of an individual cell of the specified taxa (which may be the same as a natural unit) \tab log10(μm^3^)         \cr
comments           \tab Notes from original data source                                                                 \tab -                     
}}
\source{
\url{http://www.esapubs.org/archive/ecol/E095/257/}
}
\usage{
data(Kremer_raw)
}
\description{
Kremer C.T., Gillette J.P., Rudstam L.G., Brettum P. & Ptacnik R. (2014) 
A compendium of cell and natural unit biovolumes for >1200 freshwater phytoplankton species. 
Ecology 95, 2984.
}
\details{
Ecological Archives E095-257-D1
}
\keyword{datasets}

