% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Kremer_taxa}
\alias{Kremer_taxa}
\title{Phytoplankton cell and natural unit biovolumes. 
Taxonomic look-up table based on ITIS classifications}
\format{A data frame with 2572 rows and 23 variables:

\bold{Encoding }
\encoding{UTF-8}
\tabular{lll}{
\bold{Variable} \tab \bold{Definition} \tab \bold{Storage type} \cr
original.taxa.name \tab Genus or genus + species name generated from original data source's taxa identification       \tab Character\cr
accepted.taxa.name \tab Accepted genus or genus + species name based on the ITIS categorization of original.taxa.name \tab Character\cr
original.genus     \tab Genus name from original data source's taxa identification                                    \tab Character\cr
w.tsn              \tab Working TSN from ITIS classification for given taxa                                           \tab Integer  \cr
tsn                \tab TSN matching original.taxa.name, if any                                                       \tab Integer  \cr
genus.tsn          \tab TSN matching original.genus, if any                                                           \tab Integer  \cr
a.tsn              \tab Accepted TSN reflecting taxonomic changes in ITIS to original tsn match                       \tab Integer  \cr
a.genus.tsn        \tab Accepted genus TSN reflecting taxonomic changes in ITIS to original genus.tsn match           \tab Integer  \cr
level              \tab Lowest level of taxonomic identificantion                                                     \tab Character\cr
kingdom            \tab Name of the taxa at this level                                                                \tab Character\cr
subkingdom         \tab Name of the taxa at this level                                                                \tab Character\cr
infrakingdom       \tab Name of the taxa at this level                                                                \tab Character\cr
division           \tab Name of the taxa at this level                                                                \tab Character\cr
subdivision        \tab Name of the taxa at this level                                                                \tab Character\cr
phylum             \tab Name of the taxa at this level                                                                \tab Character\cr
class              \tab Name of the taxa at this level                                                                \tab Character\cr
subclass           \tab Name of the taxa at this level                                                                \tab Character\cr
order              \tab Name of the taxa at this level                                                                \tab Character\cr
family             \tab Name of the taxa at this level                                                                \tab Character\cr
genus              \tab Name of the taxa at this level                                                                \tab Character\cr
species            \tab Name of the taxa at this level                                                                \tab Character\cr
variety            \tab Name of the taxa at this level                                                                \tab Character\cr
form               \tab Name of the taxa at this level                                                                \tab Character
}}
\source{
\url{http://www.esapubs.org/archive/ecol/E095/257/}
}
\usage{
data(Kremer_taxa)
}
\description{
Kremer C.T., Gillette J.P., Rudstam L.G., Brettum P. & Ptacnik R. (2014) 
A compendium of cell and natural unit biovolumes for >1200 freshwater phytoplankton species. 
Ecology 95, 2984.
}
\details{
Ecological Archives E095-257-D1
}
\keyword{datasets}

