% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Kremer_volumes_genus_agg}
\alias{Kremer_volumes_genus_agg}
\title{Phytoplankton cell and natural unit biovolumes.
Genus-level aggregated biovolume data and full taxonomy.}
\format{A data frame with 393 rows and 22 variables:

\bold{Encoding }
\encoding{UTF-8}
\tabular{lll}{
\bold{Variable} \tab \bold{Definition} \tab \bold{Units} \cr
accepted.taxa.name \tab Accepted genus name based on the ITIS categorization of original.taxa.name Used for aggregating biovolume data \tab -                     \cr
min_nu.biovol      \tab Taxon specific minimum biovolume                                                                               \tab log10(µm^3^)          \cr
mn_nu.biovol       \tab Taxon specific mean biovolume                                                                                  \tab log10(µm^3^)          \cr
mdn_nu.biovol      \tab Taxon specific median biovolume                                                                                \tab log10(µm^3^)          \cr
max_nu.biovol      \tab Taxon specific maximum biovolume                                                                               \tab log10(µm^3^)          \cr
sd_nu.biovol       \tab Taxon specific standard deviation of biovolume                                                                 \tab log10(µm^3^)          \cr
n                  \tab Number of biovolume estimates for given taxa                                                                   \tab -                     \cr
mn_cells.per.nu    \tab Mean number of cells per natural unit                                                                          \tab Cells per natural unit\cr
n_sources          \tab Number of distinct data sources containing biovolume estimates for given taxa                                  \tab -                     \cr
w.tsn              \tab Working TSN from ITIS classification for given taxa                                                            \tab -                     \cr
level              \tab Lowest level of identification                                                                                 \tab -                     \cr
kingdom            \tab Name of the taxa at this level                                                                                 \tab -                     \cr
subkingdom         \tab Name of the taxa at this level                                                                                 \tab -                     \cr
infrakingdom       \tab Name of the taxa at this level                                                                                 \tab -                     \cr
division           \tab Name of the taxa at this level                                                                                 \tab -                     \cr
subdivision        \tab Name of the taxa at this level                                                                                 \tab -                     \cr
phylum             \tab Name of the taxa at this level                                                                                 \tab -                     \cr
class              \tab Name of the taxa at this level                                                                                 \tab -                     \cr
subclass           \tab Name of the taxa at this level                                                                                 \tab -                     \cr
order              \tab Name of the taxa at this level                                                                                 \tab -                     \cr
family             \tab Name of the taxa at this level                                                                                 \tab -                     \cr
genus              \tab Name of the taxa at this level                                                                                 \tab -
}}
\source{
\url{http://www.esapubs.org/archive/ecol/E095/257/}
}
\usage{
data(Kremer_volumes_genus_agg)
}
\description{
Kremer C.T., Gillette J.P., Rudstam L.G., Brettum P. & Ptacnik R. (2014)
A compendium of cell and natural unit biovolumes for >1200 freshwater phytoplankton species.
Ecology 95, 2984.
}
\details{
Ecological Archives E095-257-D1
}
\keyword{datasets}

