% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Kremer_volumes_species}
\alias{Kremer_volumes_species}
\title{Phytoplankton cell and natural unit biovolumes.
Species-level biovolume data.}
\format{A data frame with 162176 rows and 12 variables:

\bold{Encoding }
\encoding{UTF-8}
\tabular{lll}{
\bold{Variable} \tab \bold{Definition} \tab \bold{Units} \cr
nu.measurement.ID  \tab Index for individual estimate of a natural unit's biovolume                                     \tab -                     \cr
data.source        \tab Source of biovolume data                                                                        \tab -                     \cr
location           \tab Sampling location as reported in original data source                                           \tab -                     \cr
original.taxa.name \tab Genus + species name generated from original data source's taxa identification                  \tab -                     \cr
accepted.taxa.name \tab Accepted genus + species name based on the ITIS categorization of original.taxa.name            \tab -                     \cr
level              \tab Level to which taxa was identified                                                              \tab -                     \cr
genus              \tab Genus name                                                                                      \tab -                     \cr
species            \tab Species name                                                                                    \tab -                     \cr
nu                 \tab Descriptor of natural unit type                                                                 \tab -                     \cr
cells.per.nu       \tab Cells per natural unit                                                                          \tab Cells per natural unit\cr
nu.biovol          \tab Biovolume corresponding to a natural unit of the specified taxa                                 \tab log10(µm^3^)          \cr
cell.biovol        \tab Biovolume of an individual cell of the specified taxa (which may be the same as a natural unit) \tab log10(µm^3^)
}}
\source{
\url{http://www.esapubs.org/archive/ecol/E095/257/}
}
\usage{
data(Kremer_volumes_species)
}
\description{
Kremer C.T., Gillette J.P., Rudstam L.G., Brettum P. & Ptacnik R. (2014)
A compendium of cell and natural unit biovolumes for >1200 freshwater phytoplankton species.
Ecology 95, 2984.
}
\details{
Ecological Archives E095-257-D1
}
\keyword{datasets}

