% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Chen_thermal_traits_references}
\alias{Chen_thermal_traits_references}
\title{Chen (2015) Thermal Traits references.}
\format{\preformatted{Classes 'tbl_df', 'tbl' and 'data.frame':	96 obs. of  1 variable:
 $ References: chr  "Baars, J. W. M. (1981) Autecological investigations on marine diatoms. 2. Generation times of 50 species. Aquatic Ecology 15, 1"| __truncated__ "Baek, S. H., S. Shimode, and T. Kikuchi. (2008) Growth of dinoflagellates, Ceratium furca and Ceratium fusus in Sagami Bay, Jap"| __truncated__ "Band-Schmidt, C. J. (2004) Effects of growth medium, temperature, salinity and seawater source on the growth of Gymnodinium cat"| __truncated__ "Barker, H. A.  (1935). The culture and physiology of the marine dinoflagellates. Arch. Mikrobiol. 6, 157-181. doi:10.1007/BF004"| __truncated__ ...
}}
\source{
\url{http://plankt.oxfordjournals.org/content/37/2/285/suppl/DC1}
}
\usage{
data(Chen_thermal_traits_references)
}
\description{
References for data from Chen B. (2015) Patterns of thermal limits of phytoplankton.
Journal of Plankton Research 37, 285–292.
}
\keyword{datasets}

