% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{Edwards_citations}
\alias{Edwards_citations}
\title{Citation information for nutrient utilization traits.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 137 rows and 2 columns.}
\source{
\url{http://www.esapubs.org/archive/ecol/E096/202/}
}
\usage{
data(Edwards_citations)
}
\description{
Data from Kyle F. Edwards, Christopher A. Klausmeier, and Elena Litchman (2015). 
Nutrient utilization traits of phytoplankton. Ecology 96:2311. 
http://dx.doi.org/10.1890/14-2252.1.
}
\details{
Table 2 of Ecological Archives E096-202-D1.
}
\keyword{datasets}

